refactor(server): merge qwen_tool_parser.py into image_gen.py
The Qwen3 XML tool-call parser is an implementation detail of image_gen's
VLM dispatch — only image_gen.py imports it, and its only purpose is
extracting `submit_edit_instruction` / `reject_request` calls from
the Qwen3.5 outputs. Same area of responsibility, fold inline.

Refs #78.
